Move StoreByBucket to its own file for testability.
