+++ searchlib/src/vespa/searchlib/docstore/CMakeLists.txt
@@ -13,6 +13,7 @@ vespa_add_library(searchlib_docstore OBJECT
     idocumentstore.cpp
     logdatastore.cpp
     logdocumentstore.cpp
+    storebybucket.cpp
     writeablefilechunk.cpp
     DEPENDS
 )
